♻️ refactor : Frame abstract 클래스로 변경
