+++ src/main/java/bowling/domain/LastFrame.java
@@ -1,26 +1,10 @@
 package bowling.domain;
 
 import bowling.global.BowlingConst;
-import java.util.Objects;
 
-public class LastFrame implements Frame {
+public class LastFrame extends Frame {
 
     private static final int LAST_FRAME_SIZE = 3;
-    private final Scores scores;
-
-    public LastFrame() {
-        this.scores = new Scores();
-    }
-
-    @Override
-    public void addScore(Score score) {
-        if (isRemainChance()) {
-            this.scores.add(score);
-            ScoreValidator.validate(this);
-            return;
-        }
-        throw new IllegalArgumentException("남은 기회가 없습니다.");
-    }
 
     @Override
     public boolean isRemainChance() {
@@ -40,27 +24,5 @@ public class LastFrame implements Frame {
             return true;
         }
         return false;
-    }
-
-    @Override
-    public Scores scores() {
-        return this.scores;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        LastFrame lastFrame = (LastFrame) o;
-        return Objects.equals(this.scores, lastFrame.scores);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(this.scores);
     }
 }
